pyproject.toml
[tool.poetry]
name = "n8n-agents"
version = "0.1.0"
description = "n8n prototype to connect ai agents to django"
authors = ["Camilo Monsalve Gómez <beren.monsalve@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "4.2.21"
python-dotenv = "1.0.1"
djangorestframework = "3.15.2"
django-extensions = "3.2.3"
psycopg2 = "^2.9.10"
django-filter = "24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
